# Julia sample bot