--- a/GameEngine/Battleships/Battleships/App.config
+++ b/GameEngine/Battleships/Battleships/App.config
@@ -22,6 +22,9 @@
             <setting name="PathToJava" serializeAs="String">
                 <value>java</value>
             </setting>
+            <setting name="PathToJulia" serializeAs="String">
+                <value>julia</value>
+            </setting>
             <setting name="PathToGolang" serializeAs="String">
                 <value>go</value>
             </setting>