--- a/GameEngine/Battleships/Battleships/Properties/Settings.settings
+++ b/GameEngine/Battleships/Battleships/Properties/Settings.settings
@@ -14,6 +14,9 @@
     <Setting Name="PathToJava" Type="System.String" Scope="Application">
       <Value Profile="(Default)">java</Value>
     </Setting>
+    <Setting Name="PathToJulia" Type="System.String" Scope="Application">
+      <Value Profile="(Default)">julia</Value>
+    </Setting>
     <Setting Name="PathToGolang" Type="System.String" Scope="Application">
       <Value Profile="(Default)">go</Value>
     </Setting>