--- a/GameEngine/Battleships/BotRunner/BotRunner.csproj
+++ b/GameEngine/Battleships/BotRunner/BotRunner.csproj
@@ -54,6 +54,7 @@
     <Compile Include="Harness\Bot\Compilers\GolangCompiler.cs" />
     <Compile Include="Harness\Bot\Compilers\JavaCompiler.cs" />
     <Compile Include="Harness\Bot\Compilers\JavaScriptCompiler.cs" />
+    <Compile Include="Harness\Bot\Compilers\JuliaCompiler.cs" />
     <Compile Include="Harness\Bot\Compilers\PythonCompiler.cs" />
     <Compile Include="Harness\Bot\ICompiler.cs" />
     <Compile Include="Harness\Bot\StringToPlaceShipCommand.cs" />
@@ -61,6 +62,7 @@
     <Compile Include="Harness\Bot\Runners\GolangRunner.cs" />
     <Compile Include="Harness\Bot\Runners\JavaRunner.cs" />
     <Compile Include="Harness\Bot\Runners\JavaScriptRunner.cs" />
+    <Compile Include="Harness\Bot\Runners\JuliaRunner.cs" />
     <Compile Include="Harness\Bot\Runners\PythonRunner.cs" />
     <Compile Include="Harness\ConsoleHarness\ConsoleHarness.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />