--- a/GameEngine/Battleships/BotRunner/Harness/Bot/BotCompiler.cs
+++ b/GameEngine/Battleships/BotRunner/Harness/Bot/BotCompiler.cs
@@ -29,6 +29,8 @@ namespace TestHarness.TestHarnesses.Bot
                     break;
                     case BotMeta.BotTypes.Java:
                     _compiler = new JavaCompiler(botMeta, botDir, compileLogger);
+                    case BotMeta.BotTypes.Julia:
+                    _compiler = new JuliaCompiler(botMeta, botDir, compileLogger);
                     break;
                     case BotMeta.BotTypes.Python2:
                     case BotMeta.BotTypes.Python3: