--- a/GameEngine/Battleships/BotRunner/Harness/Bot/BotHarness.cs
+++ b/GameEngine/Battleships/BotRunner/Harness/Bot/BotHarness.cs
@@ -63,6 +63,9 @@ namespace TestHarness.TestHarnesses.Bot
                 case BotMeta.BotTypes.JavaScript:
                     _botRunner = new JavaScriptRunner(this);
                     break;
+                case BotMeta.BotTypes.Julia:
+                    _botRunner = new JuliaRunner(this);
+                    break;
                 default:
                     throw new ArgumentException("Invalid bot type " + meta.BotType);
             }