--- a/GameEngine/Battleships/BotRunner/Harness/Bot/Compilers/JuliaCompiler.cs
+++ b/GameEngine/Battleships/BotRunner/Harness/Bot/Compilers/JuliaCompiler.cs
@@ -1,0 +1,50 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+using BotRunner.Properties;
+using BotRunner.Util;
+using Domain.Bot;
+using Domain.Meta;
+using GameEngine.Loggers;
+
+namespace TestHarness.TestHarnesses.Bot.Compilers
+{
+    public class JuliaCompiler : ICompiler
+    {
+        private readonly BotMeta _botMeta;
+        private readonly string _botDir;
+        private readonly ILogger _compileLogger;
+        private bool _failed;
+
+        public JuliaCompiler(BotMeta botMeta, string botDir, ILogger compileLogger)
+        {
+            _botMeta = botMeta;
+            _botDir = botDir;
+            _compileLogger = compileLogger;
+        }
+
+        public bool HasPackageManager()
+        {
+            return false; // package manager is built in
+        }
+
+        public bool RunPackageManager()
+        {
+            return true; // the compile script can install packages if it must.
+        }
+
+        public bool RunCompiler()
+        {
+            // don't actully compile anything for now
+            _compileLogger.LogInfo("Compiling bot " + _botMeta.NickName + " using Julia");
+            return true;
+        }
+
+        void ProcessDataRecieved(object sender, System.Diagnostics.DataReceivedEventArgs e)
+        {
+            _compileLogger.LogInfo(e.Data);
+        }
+    }
+}