--- a/GameEngine/Battleships/BotRunner/Properties/Settings.settings
+++ b/GameEngine/Battleships/BotRunner/Properties/Settings.settings
@@ -15,7 +15,7 @@
       <Value Profile="(Default)">java</Value>
     </Setting>
     <Setting Name="PathToJulia" Type="System.String" Scope="User">
-      <Value Profile="(Default)">java</Value>
+      <Value Profile="(Default)">julia</Value>
     </Setting>
     <Setting Name="PathToGolang" Type="System.String" Scope="User">
       <Value Profile="(Default)">go</Value>