--- a/GameEngine/Battleships/Domain/Bot/BotMeta.cs
+++ b/GameEngine/Battleships/Domain/Bot/BotMeta.cs
@@ -16,12 +16,13 @@ namespace Domain.Bot
         {
             CSharp,
             Java,
-            JavaScript,
+            JavaScript,            
             CPlusPlus,
             Python2,
             Python3,
             FSharp,
-            Golang
+            Golang,
+            Julia
         }
     }
 }