--- a/Sample Bots/Julia/README.md
+++ b/Sample Bots/Julia/README.md
@@ -1,0 +1,9 @@
+## Julia
+
+sampleBot
+
+Just uses julia standard library
+
+### Installing Julia
+
+See https://julialang.org/downloads/