--- a/Sample Bots/Julia/main.jl
+++ b/Sample Bots/Julia/main.jl
@@ -1,0 +1,1 @@
+print("Hello Julia")