--- a/Sample Bots/Julia/main.jl
+++ b/Sample Bots/Julia/main.jl
@@ -2,8 +2,8 @@ import JSON
 
 commandFileName = "command.txt"
 placeShipFileName = "place.txt"
-#stateFileName = "test_data/Phase 1 - Round 0/state.json"
-stateFileName = "test_data/Phase 2 - Round 1/state.json"
+stateFileName = "test_data/Phase 1 - Round 0/state.json"
+#stateFileName = "test_data/Phase 2 - Round 1/state.json"
 #stateFileName = "state.json"
 
 state = JSON.parsefile(stateFileName)
@@ -16,7 +16,6 @@ function makeYourMove()
   if phase == 1
     deployShips()
   elseif phase == 2
-    println("Bombs away!..")
     bombsAway()
   else
     throw(Exception("Unsupported phase: '$phase'"))
@@ -25,6 +24,17 @@ end
 
 function deployShips()
   println("* Deploying ships..")
+
+  placementString = """
+Carrier 0 0 North
+Battleship 1 0 North
+Cruiser 2 0 North
+Submarine 3 0 North
+Destroyer 4 0 North
+""";
+
+	 write(placeShipFileName, placementString)
+   println("* Aight Captain, the ships were deployed \o/");
 end
 
 function bombsAway()