--- a/Sample Bots/Julia/main.jl
+++ b/Sample Bots/Julia/main.jl
@@ -88,3 +88,6 @@ end
 
 println("ARGS=$ARGS")
 main(ARGS...)
+
+#TODO: facture out bot module
+#TODO: write some basic tests..