--- a/Sample Bots/Julia/main.jl
+++ b/Sample Bots/Julia/main.jl
@@ -1,26 +1,38 @@
 import JSON
 
-commandFileName = "command.txt"
-placeShipFileName = "place.txt"
-#stateFileName = "test_data/Phase 1 - Round 0/state.json"
-#stateFileName = "test_data/Phase 2 - Round 1/state.json"
-stateFileName = "state.json"
-playerKey = "007"
-workingDirectory = "."
+const commandFileName = "command.txt"
+const placeShipFileName = "place.txt"
+const stateFileName = "state.json"
 
-function makeYourMove(state)
-  phase = state["Phase"]
+type GameContext
+  state::Dict
+  playerKey::String
+  workingDirectory::String
+  GameContext()=new(Dict(), "", "")
+end
+
+function getFilePath(gameContext::GameContext, fileName::AbstractString)
+  ret = string(gameContext.workingDirectory, '/', fileName)
+  println("** $ret")
+  ret
+end
+function getPhase(gameContext::GameContext)
+  gameContext.state["Phase"]
+end
+
+function makeYourMove(gameContext::GameContext)
+  phase = getPhase(gameContext)
   println("* Phase $phase")
   if phase == 1
-    deployShips(state)
+    deployShips(gameContext)
   elseif phase == 2
-    bombsAway(state)
+    bombsAway(gameContext)
   else
-    throw(Exception("Unsupported phase: '$phase'"))
+    error("Unsupported phase: '$phase'")
   end
 end
 
-function deployShips(state)
+function deployShips(gameContext::GameContext)
   println("* Deploying ships..")
 
   placementString = """
@@ -31,11 +43,11 @@ Submarine 3 0 North
 Destroyer 4 0 North
 """
 
-	 write(placeShipFileName, placementString)
+	 write(getFilePath(gameContext, placeShipFileName), placementString)
    println("* Aye aye Captain, the ships were deployed \\o/")
 end
 
-function bombsAway(state)
+function bombsAway(gameContext::GameContext)
   println("* Bombs away!")
 
   fire = 1
@@ -44,33 +56,33 @@ function bombsAway(state)
 
   payload = string(fire, ",", xCoordinate, ",", yCoordinate, "\n")
 
-  write(commandFileName, payload)
+  write(getFilePath(gameContext, commandFileName), payload)
   println("* Shot fired: [$xCoordinate,$yCoordinate]")
 end
 
 function run(playerKey, workPathArray...)
   #I'm not sure why args containing strings are split :(
-  workingDirectory = join(workPathArray, " ")
+  gameContext = GameContext()
+  gameContext.playerKey = playerKey
+  gameContext.workingDirectory = join(workPathArray, " ")
   println("* playerKey:        '$playerKey'")
-  println("* workingDirectory: '$workingDirectory'")
-  if !isdir(workingDirectory)
+  println("* workingDirectory: '$(gameContext.workingDirectory)'")
+  if !isdir(gameContext.workingDirectory)
       printUsage()
-      println("Error: Working directory \"$workingDirectory\" does not exist.")
+      error("Error: Working directory \"$(gameContext.workingDirectory)\" does not exist.")
       exit(1)
   end
 
-  qstateFileName = string(workingDirectory, '/', stateFileName)
-  println(qstateFileName)
-  state = JSON.parsefile(qstateFileName)
+  gameContext.state = JSON.parsefile(getFilePath(gameContext, stateFileName))
   #print(state)
-  makeYourMove(state)
+  makeYourMove(gameContext)
 end
 
 function main(args...)
   println("args=$args")
   if length(args) < 2
       printUsage()
-      println("Error: Expected 2 args but got: $args")
+      error("Error: Expected 2 args but got: $args")
       exit(1)
   end
   run(args...)