… Sample Bots/Julia/bot.json
… Sample Bots/Julia/test_data/Phase 1 - Round 0/state.json
… Sample Bots/Julia/test_data/Phase 2 - Round 1/state.json