Hi,
I only tested it on linux sofar, a colleague said he'll try to test it on windows but I thought I'll give you the heads up solong.

Also since it is a just in time compiled language the startup times vary quite a bit at the moment (1.5-3s for the random bot). I'm still looking into statically compiling it but could not get it right yet. So I don't think it is really feasible to write julia bot's yet.