@EntelectChallenge 2017-05-17
Hi Amanica,

With sample bot submission closing tomorrow night, would you like us to include this as one of the supported languages?  You will have to risk the random startup times however and it does make it a bit more difficult to work with in terms of the calibration phase.