@amanica 2017-05-17
hi,
Oh, on the sample bot readme it says "Sample bot submissions will close at
Midnight on the 26th of May"
https://github.com/EntelectChallenge/2017-Battleships/tree/master/Sample%20Bots
So I thought I have a bit more time.

I'd obviously like for it to be included, but I'd probably need to reach
out to the julia community to help me statically compile it (or to find out
if it is possible). I'll reach out to them now, maybe I'll have an answer
by tomorrow.

thanks
Marius Kruger




On 17 May 2017 at 20:32, Entelect Challenge <notifications@github.com>
wrote:

> Hi Amanica,
>
> With sample bot submission closing tomorrow night, would you like us to
> include this as one of the supported languages? You will have to risk the
> random startup times however and it does make it a bit more difficult to
> work with in terms of the calibration phase.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/EntelectChallenge/2017-Battleships/pull/37#issuecomment-302187568>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAz2qO4fb0ASi7ubMI-Qjf5W5mcCsQuaks5r6z04gaJpZM4Nagyh>
> .
>



-- 
✝ Marius