@amanica 2017-05-17
I managed to 'compile' it into an 'exe' now but that didn't really make it
faster or more consistent
(This isn't static compile yet unfortunately).

From the messy stats below, i deduce that about 1 in 15 runs (compiled or
not) goes 2s over the average
and this is just placing the ships in hard-coded places..

The only way I can see this working at this point is if the game engine
will allow re-running a move if it goes over time and if it continues to be
over the limit only then penalise it. (but then the working dir needs to be
reset to how it was before the move to not allowing caching between tries)

:'(

==
julia> function stats(txt); A = [parse(Float64,x) for x in split(txt, "
")]; avg = sum(A)/length(A); B = max(0.0, A - avg - 2) ;
println("avg=$avg") ; println(A); println(B); end

  1.035582 seconds (288.30 k allocations: 12.316 MB, 6.38% gc time)
raw times: 2 1.99 2.04 2.06 2.06 2.05 2.05 3 2.03 2.03 2.36 1.94 3.47 *6.79*
4.38
1 loops, best of 15: 1.94 sec per loop
amanica@amanica-laptop:/stf/prj/2017entelectchallenge/2017-Battleships/Sample
Bots/Julia (julia-sample-bot)>
$ python -m timeit -n 1 -r 15 -v -s 'import os' 'os.system("julia --
main.jl 007 test_data/Phase\ 1\ -\ Round\ 0/")'

julia> stats("2 1.99 2.04 2.06 2.06 2.05 2.05 3 2.03 2.03 2.36 1.94 3.47
6.79 4.38")
avg=2.683333333333334
[2.0,1.99,2.04,2.06,2.06,2.05,2.05,3.0,2.03,2.03,2.36,1.94,3.47,*6.79*,4.38]
[0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,2.10667,0.0]



  0.985178 seconds (287.32 k allocations: 12.517 MB, 6.63% gc time)
raw times: 1.78 4.09 3.07 3.52 3.86 3.5 4.05 4.09 3.44 3.22 3.52 7.13 *5.5*
3.86 3.86
1 loops, best of 15: 1.78 sec per loop
amanica@amanica-laptop:/stf/prj/2017entelectchallenge/2017-Battleships/Sample
Bots/Julia (julia-sample-bot)>
$ python -m timeit -n 1 -r 15 -v -s 'import os' 'os.system("target/bot.exe
007 test_data/Phase\ 1\ -\ Round\ 0/")'

julia> stats("1.78 4.09 3.07 3.52 3.86 3.5 4.05 4.09 3.44 3.22 3.52 7.13
5.5 3.86 3.86")
avg=3.8993333333333338
[1.78,4.09,3.07,3.52,3.86,3.5,4.05,4.09,3.44,3.22,3.52,7.13,*5.5*,3.86,3.86]
[0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,1.23067,0.0,0.0,0.0]

On 17 May 2017 at 20:32, Entelect Challenge <notifications@github.com>
wrote:

> Hi Amanica,
>
> With sample bot submission closing tomorrow night, would you like us to
> include this as one of the supported languages? You will have to risk the
> random startup times however and it does make it a bit more difficult to
> work with in terms of the calibration phase.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/EntelectChallenge/2017-Battleships/pull/37#issuecomment-302187568>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAz2qO4fb0ASi7ubMI-Qjf5W5mcCsQuaks5r6z04gaJpZM4Nagyh>
> .
>



-- 
✝ Marius