@EntelectChallenge 2017-05-18
Hi,

I see now that I confused the dates.  You are correct, it only closes on the 26th so you have some more time to play around with this to see if you can improve it?