@amanica 2017-05-18
awesome yes I'll try to fix it..

On 18 May 2017 7:21 AM, "Entelect Challenge" <notifications@github.com>
wrote:

> Hi,
>
> I see now that I confused the dates. You are correct, it only closes on
> the 26th so you have some more time to play around with this to see if you
> can improve it?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/EntelectChallenge/2017-Battleships/pull/37#issuecomment-302303412>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAz2qH9Llzs07A1xjP3yQu8O3wjtFQIeks5r69VcgaJpZM4Nagyh>
> .
>