@amanica 2017-05-26
After adding some command-line options, it seems to run much more consistently!  
So please include it.

---
New commandline options: --precompiled=yes --compilecache=no
stats:
average_run_time: 2.8456 
average difference with average_run_time: 0.03516800000000009
max difference with average_run_time: 0.13439999999999985

i.e. in 25 runs the max offset to the average is 0.134 seconds which I think is totally acceptable..


---
how I calculated my stats if anybody is interested:

python -m timeit -n 1 -r 25 -v -s 'import os' 'os.system("julia --precompiled=yes --compilecache=no -- main.jl 007 test_data/Phase\ 2\ -\ Round\ 1/")'
  0.582302 seconds (241.16 k allocations: 10.284 MB)
raw times: 2.84 2.84 2.84 2.82 2.81 2.82 2.83 2.85 2.91 2.85 2.82 2.81 2.81 2.83 2.9 2.94 2.98 2.86 2.78 2.87 2.89 2.82 2.77 2.84 2.81
1 loops, best of 25: 2.77 sec per loop

julia> function stats(txt); A = [parse(Float64,x) for x in split(txt, " ")]; avg = mean(A); B = max(0.0, A - avg - 2); C = abs(A - avg) ; println("avg=$avg avg_delta=$(mean(C)) max_delta=$(max(C...))") ; println(A); println(C); println(B); end

julia> stats("2.84 2.84 2.84 2.82 2.81 2.82 2.83 2.85 2.91 2.85 2.82 2.81 2.81 2.83 2.9 2.94 2.98 2.86 2.78 2.87 2.89 2.82 2.77 2.84 2.81")
avg=2.8456 avg_delta=0.03516800000000009 max_delta=0.13439999999999985
[2.84,2.84,2.84,2.82,2.81,2.82,2.83,2.85,2.91,2.85,2.82,2.81,2.81,2.83,2.9,2.94,2.98,2.86,2.78,2.87,2.89,2.82,2.77,2.84,2.81]
**[0.0056,0.0056,0.0056,0.0256,0.0356,0.0256,0.0156,0.0044,0.0644,0.0044,0.0256,0.0356,0.0356,0.0156,0.0544,0.0944,0.1344,0.0144,0.0656,0.0244,0.0444,0.0256,0.0756,0.0056,0.0356]**
[0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0]